requirements.txt

streamlit>=1.28.0
requests>=2.31.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
transformers[torch]>=4.35.0